move standalone tests into a dedicated package
